@@ -1,17 +1,18 @@
 package org.biologer.biologer;
 
-import android.app.Activity;
 import android.content.Intent;
 import android.net.Uri;
 import android.os.Bundle;
-import android.support.v4.app.FragmentTransaction;
+
+import androidx.fragment.app.Fragment;
+
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.TextView;
 
 
-public class AboutFragment extends android.support.v4.app.Fragment {
+public class AboutFragment extends Fragment {
 
     TextView tv_database;
     TextView address;
