@@ -3,18 +3,11 @@ package org.biologer.biologer;
 import android.app.NotificationChannel;
 import android.app.NotificationManager;
 import android.os.Build;
-import android.support.multidex.MultiDexApplication;
-import android.util.Log;
+import androidx.multidex.MultiDexApplication;
 
 import org.biologer.biologer.model.DaoMaster;
 import org.biologer.biologer.model.DaoSession;
-import org.biologer.biologer.model.Taxon;
-import org.biologer.biologer.model.TaxonLocalization;
-import org.biologer.biologer.model.TaxonLocalizationDao;
 import org.greenrobot.greendao.database.Database;
-
-import java.util.List;
-import java.util.Locale;
 
 /**
  * Created by brjovanovic on 12/24/2017.
