@@ -1,8 +1,8 @@
 package org.biologer.biologer;
 
 import android.content.Context;
-import android.content.SharedPreferences;
-import android.support.v7.preference.PreferenceManager;
+
+import androidx.preference.PreferenceManager;
 import android.util.Log;
 
 import org.biologer.biologer.model.DaoMaster;
